Update:
Collision:

collision has been all but perfected. But the code is too long, I haven't figured out a shorter way to do it. It took me a long time to just simply make it work.

The way it works is that because I simply could not find a way to prevent the player from entering the wall by a few pixels, the if-statement checks if the player has entered a wall, then pushes it back by the number of pixels it entered.

Dice movement:

Dice movement works flawlessly for as much as I tested it. Problem is I can't test every weird someone could control their character.

 If the user plays optimally, moving square to square then simply setting speed to zero when player runs out of dice should be enough.

  However if the player moves in a weird way the game would stop while the player character is not fully in a tile block, there rest of the code works to snap the character back, or further into the block in which most of it is in.

Signed-off-by: Abdulmoyn Sommakieh <moyn@archiqoo.com>